src/pyproject.toml
[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "mcpo-simple-server"
dynamic = ["version"]
description = "Python-based LLM server that implements the Model Context Protocol (MCP)"
readme = "mcpo_simple_server/README.md"
authors = [
    {name = "Artur Zdolinski", email = "contact@nchekwa.com"}
]
license = {text = "MIT"}
classifiers = [
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.103.1",
    "uvicorn>=0.23.2",
    "pydantic>=2.3.0",
    "sse-starlette>=1.6.5",
    "loguru>=0.7.0",
    "python-dotenv>=1.0.0",
    "uuid>=1.30",
    "python-jose[cryptography]",
    "passlib[bcrypt]>=1.7.4",
    "aiofiles>=23.1.1",
]

[project.urls]
"Homepage" = "https://getsimpletool.org"
"Bug Tracker" = "https://github.com/getsimpletool/mcpo-simple-server/issues"

[project.scripts]
mcpo-simple-server = "mcpo_simple_server.__main__:main"

[tool.setuptools]
packages = ["mcpo_simple_server"]

[tool.setuptools.package-data]
mcpo_simple_server = ["**/*.md", "**/*.sh"]

[tool.setuptools.exclude-package-data]
mcpo_simple_server = ["data/**", ".env"]

[tool.setuptools.dynamic]
version = {attr = "mcpo_simple_server._version.__version__"}
